Uploaded Files to Repository
Server and Client are to be run together.
Text documents contains the user information and the event information. 
User information can be deleted and it will create itself during runtime.
Event information can be edited, but is needed to run. Word document contains information about the system and how it works.